# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mortgage Calculator API",
        "description": "REST API for mortgage simulation using the French amortization method. Supports both fixed and variable interest rates, with complete simulation history and export functionality.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/simulations": {
            "get": {
                "tags": [
                    "Simulation Management"
                ],
                "summary": "List user simulations",
                "description": "Retrieve a paginated list of all simulations for the authenticated user, ordered by most recent first.",
                "operationId": "010213609cfe3000816f2ab33723efcb",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Simulations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "loan_amount": {
                                                        "type": "string",
                                                        "example": "200000.00"
                                                    },
                                                    "duration_months": {
                                                        "type": "integer",
                                                        "example": 360
                                                    },
                                                    "rate_type": {
                                                        "type": "string",
                                                        "example": "fixed"
                                                    },
                                                    "annual_rate": {
                                                        "type": "string",
                                                        "example": "3.50"
                                                    },
                                                    "monthly_payment": {
                                                        "type": "string",
                                                        "example": "898.09"
                                                    },
                                                    "total_amount": {
                                                        "type": "string",
                                                        "example": "323312.40"
                                                    },
                                                    "total_interest": {
                                                        "type": "string",
                                                        "example": "123312.40"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-11-03T10:30:00+00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/simulations?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/simulations?page=2"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/simulations?page=2",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token"
            }
        }
    },
...
}